Move SubscriptionTableController to resource package so tests with API profile can discover it during context load
