+++ src/main/java/org/candlepin/subscriptions/resource/SubscriptionResource.java
@@ -20,18 +20,16 @@
  */
 package org.candlepin.subscriptions.resource;
 
-import java.time.OffsetDateTime;
-import javax.validation.constraints.Min;
 import org.candlepin.subscriptions.security.auth.ReportingAccessRequired;
-import org.candlepin.subscriptions.subscription.SubscriptionTableController;
 import org.candlepin.subscriptions.utilization.api.model.*;
 import org.candlepin.subscriptions.utilization.api.resources.SubscriptionsApi;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
+
+import javax.validation.constraints.Min;
+import java.time.OffsetDateTime;
 
 /** Subscriptions Table API implementation. */
 @Component
-@Profile("capacity-ingress")
 public class SubscriptionResource implements SubscriptionsApi {
 
   private final SubscriptionTableController subscriptionTableController;
